project/Sprint6-1/requirements.txt
aws-cdk-lib==2.17.0
constructs>=10.0.0,<11.0.0
aws-cdk.core>=1.135.0
aws-cdk.assertions>=1.135.0